Wrap login token request call on server side
It allows to hook on requestLoginTokenForUser.

The Meteor.call('requestLoginTokenForUser', options); on the server side is required because requestLoginTokenForUser is registered as a Meteor method in the accounts-passwordless package.
As stated in the Meteor.call documentation (https://docs.meteor.com/api/methods.html#Meteor-call), the call is performed synchronously as no callback is provided.